+++ terraform/modules/iam/variables.tf
@@ -1,3 +1,8 @@
+variable "dynamodb-table-banshee-44-mods-backend-mods-arn" {
+  description = "banshee-44-mods-backend-mods DynamoDB table ARN"
+  type = string
+}
+
 variable "error-sns-topic" {
   description = "SNS Topic ARN to trigger on lambda failure"
   type = string
@@ -5,10 +10,5 @@ variable "error-sns-topic" {
 
 variable "parameter-store-twitter-auth-arn" {
   description = "Parameter Store twitter auth config path ARN"
-  type = string
-}
-
-variable "aws-dynamodb-table-banshee-44-mods-bot-arn" {
-  description = "DynamoDB table ARN"
   type = string
 }
