+++ terraform/variables.tf
@@ -4,6 +4,11 @@ variable "aws_region" {
   default = "us-east-1"
 }
 
+variable "dynamodb-table-banshee-44-mods-backend-mods-arn" {
+  description = "banshee-44-mods-backend-mods DynamoDB table ARN"
+  type = string
+}
+
 variable "error-sns-topic" {
   description = "SNS Topic ARN to trigger on lambda failure"
   type = string
